refactor: changes for a more accurate code
