--- src/Commands.ts
@@ -1,8 +1,0 @@
-import { Command } from './Command';
-import { FxTweet } from './commands/FxTweet';
-import { Hello } from './commands/Hello';
-
-export const Commands: Command[] = [
-    Hello, 
-    FxTweet
-];
