+++ src/commands/Hello.ts
@@ -1,15 +1,24 @@
-import { CommandInteraction, Client, ApplicationCommandType } from 'discord.js';
+import {
+  CommandInteraction,
+  Client,
+  ApplicationCommandType,
+  InteractionResponse,
+} from 'discord.js';
+
 import { Command } from '../Command';
 
 export const Hello: Command = {
   name: 'hello',
   description: 'Returns a greeting',
   type: ApplicationCommandType.ChatInput,
-  run: async (client: Client, interaction: CommandInteraction) => {
+  run: (
+    client: Client,
+    interaction: CommandInteraction,
+  ): Promise<InteractionResponse<boolean>> => {
     const username = interaction.user?.username;
     const content = `Ara ara... Hi, ${username}-kun!`;
 
-    await interaction.reply({
+    return interaction.reply({
       ephemeral: true,
       content,
     });
