+++ src/interfaces/Commands.ts
@@ -1,0 +1,7 @@
+import { FxTweet } from '../commands/FxTweet';
+import { Hello } from '../commands/Hello';
+
+export const Commands = [
+    Hello, 
+    FxTweet
+];
