+++ src/listeners/interactionCreate.ts
@@ -1,5 +1,5 @@
 import { Client, CommandInteraction, Interaction } from 'discord.js';
-import { Commands } from '../Commands';
+import { Commands } from '../interfaces/Commands';
 
 export default (client: Client): void => {
   client.on('interactionCreate', async (interaction: Interaction) => {
@@ -13,7 +13,9 @@ const handleSlashCommand = async (
   client: Client,
   interaction: CommandInteraction,
 ): Promise<void> => {
-  const slashCommand = Commands.find(command => command.name === interaction.commandName);
+  const slashCommand = Commands.find(
+    command => command.name === interaction.commandName,
+  );
   if (!slashCommand) {
     interaction.followUp({ content: 'An error has occurred' });
     return;
